+++ NEWS.md
@@ -3,6 +3,7 @@
 - Family common names now included in eBird taxonomy
 - `auk_select()` now requires certain columns to be kept
 - Better handling of file paths with `prefix` argument in `auk_split()`
+- Fixed bug causing undescribed species to be dropped by `auk_rollup()`
 
 # auk 0.4.0